improve deploy of libs and deploy for all bins
now the script checks the dependencies of existing libs in the AppDir before deploying, as before the script would only check the deps of the libs it deployed. 

it will also deploy the libs of all bins and not just `$TARGET`, the count of deployed libs has also been simplified as well as the warning for not found libraries.